Refactored importer_app with importer name
