+++ concordia/models.py
@@ -8,8 +8,6 @@ from django.contrib.postgres.fields import JSONField
 from django.db import models
 from django_prometheus_metrics.models import MetricsModelMixin
 
-from importer.importer.tasks import download_async_collection
-
 metadata_default = dict
