+++ importer/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class ImporterAppConfig(AppConfig):
-    name = 'importer_app'
+    name = 'importer'
